requirements.txt
python-dotenv
pypdf
langchain
PyPDF2==3.0.1
tiktoken==0.8.0
langchain-text-splitters==0.3.3
chromadb==0.5.0
sentence-transformers==2.7.0
google-generativeai==0.8.3
python-dotenv==1.0.1
streamlit==1.33.0
